Comments and changes to Movie and Controller
Added comments to all classes and added functions to Movie and
Controller classes. In the Movie class I added Functions to add new
DVD's and Actors to the movie. I also created the interface Searchable
that the Movie class implements to allow for easy searches of Movies. I
added getters and setters to the Movie and DVD classes.  The Payment
class has been added and the Gender enum has been moved to its own file.