Add review functionality and more
Added functions to allow for reviews to be created and added to Movies
and Customers. Created a function to get and calculate the rating for a
Movie and store it in the Movie rating variable. Created functions to
search through the customers and movies list in the Controller class.
Customer now implements the Searchable interface to allow for searching.
Added getters to the Review class. Added more function calls in the Main
class to add Customers to the Controllers customers list.