Made changes to Controller, MovieRental, and Rental
In the Controller I made it so it actually removes the rental from the
rentals list now instead of showing it in the list as AVAILABLE. I
updated the calls for returnRental in MovieRental. I also added
getStatus in Rental as well as had it check for the DVD serial number in
the contains function.