More updates and comments
In Controller I added requests to printLists, started working on
payments with lateFee and makePayment, and added requestAnswered. Added
getStatus and setStatus to DVD. Made a few minor changes to Movie. In
MovieRental I updated the addDVD and addRental functions, and added
requests as well as requests that had been answered. Added
getPaymentMethod to Payment. Added getDvd to Rental. Added info and
contains to Request